requirements.txt
streamlit>=1.29.0
google-generativeai>=0.3.2
chromadb>=0.4.22
sqlalchemy>=2.0.23
pandas>=2.1.4
numpy>=1.26.0
pillow>=10.1.0
python-dotenv>=1.0.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6 